requirements.txt
# Day 18
colorgram.py
# Day 25
pandas
# Day 29
pyperclip

# Day 32
python-dotenv  # not part of course content

# Day 33
requests

# Day 25
twilio

# Day 45
beautifulsoup4

# Day 46
spotipy